DEV-358: Change dna nanograms from double to int to enforce rounding

+++ patient-db/src/main/java/com/hartwig/hmftools/patientdb/data/SampleData.java
@@ -21,7 +21,7 @@ public abstract class SampleData {
     public abstract LocalDate samplingDate();
 
     @Nullable
-    public abstract Double dnaNanograms();
+    public abstract Integer dnaNanograms();
 
     @Nullable
     public abstract Double tumorPercentage();
