+++ patient-db/src/main/resources/patches/patientdb2.1_to_2.2_migration.sql
@@ -1,2 +1,2 @@
 ALTER TABLE sample
-    ADD COLUMN dnaNanograms DOUBLE PRECISION AFTER samplingDate;
+    ADD COLUMN dnaNanograms int AFTER samplingDate;
